requirements.txt
# ------------------------
# Core Packages
# ------------------------
pyspark>=3.5.0,<4.0            # Apache Spark for local execution
pandas>=2.1.0,<3.0             # Data manipulation and processing
pyarrow>=12.0,<13.0            # Parquet/Arrow support for Spark and DBT

# ------------------------
# DBT for Transformations
# ------------------------
dbt-core>=1.10.0,<1.11          # Core DBT functionality
dbt-databricks>=1.10.0,<1.11    # DBT adapter for Databricks
dbt-spark>=1.10.0,<1.11         # DBT adapter for Spark (for local testing)

# ------------------------
# File Download / Extraction
# ------------------------
requests>=2.31.0,<3.0           # For downloading files via HTTP
tqdm>=4.65.0,<5.0               # Progress bars for downloads
zipfile36>=0.1.3                # Extract ZIP files across Python versions

# ------------------------
# Airflow for Orchestration
# ------------------------
apache-airflow>=2.7.0,<3.0                              # Core Airflow package
apache-airflow-providers-apache-spark>=3.5.0,<4.0      # Spark provider for Airflow
apache-airflow-providers-databricks>=5.1.0,<6.0        # Databricks provider for Airflow

# ------------------------
# Utilities
# ------------------------
python-dotenv>=1.0.0,<2.0       # Load environment variables from .env files
